requirements.txt
Flask
Flask-CORS
pandas
requests
python-dotenv
gunicorn
fuzzywuzzy
scikit-learn
numpy